Implement Bond CLI with with init functionality
Creates a Bond CLI app. This app will be extended to allow transactions
to be submitted to the Bond REST API.

This commit implements one command, `init`, which takes a URL and
stores it in a config file. This will file will be read by other
commands when submitting transactions.

Signed-off-by: Darian Plumb <dplumb@bitwise.io>